impl Sub<CsMat> and Mul<N> traits for CsMat. needs tests
