+++ src/sparse/compressed.rs
@@ -2,9 +2,7 @@
 
 
 use sparse::csmat::{CsMat, CsMatVec, CsMatView};
-use sparse::binop;
-use num::traits::Num;
-use std::ops::{Deref, Add};
+use std::ops::{Deref};
 
 /// The SpMatView trait describes data that can be seen as a view
 /// into a CsMat
@@ -25,19 +23,3 @@ where N: Copy,
     }
 }
 
-impl<'a, 'b, N, IStorage, DStorage, Mat> Add<&'b Mat>
-for &'a CsMat<N, IStorage, DStorage>
-where N: 'a + Copy + Num + Default,
-      IStorage: 'a + Deref<Target=[usize]>,
-      DStorage: 'a + Deref<Target=[N]>,
-      Mat: SpMatView<N> {
-    type Output = CsMatVec<N>;
-
-    fn add(self, rhs: &'b Mat) -> CsMatVec<N> {
-        if self.storage() != rhs.borrowed().storage() {
-            return binop::add_mat_same_storage(
-                self, &rhs.borrowed().to_other_storage()).unwrap()
-        }
-        binop::add_mat_same_storage(self, rhs).unwrap()
-    }
-}
